+++ android/src/main/kotlin/cl/ceisufro/fluttermbtilesextractor/ExecutionCallback.kt
@@ -1,0 +1,5 @@
+package cl.ceisufro.fluttermbtilesextractor
+
+interface ExecutionCallback {
+    fun onTaskFinish(extractResult: ExtractResult?)
+}