+++ android/src/main/kotlin/cl/ceisufro/fluttermbtilesextractor/Utils.kt
@@ -1,0 +1,18 @@
+package cl.ceisufro.fluttermbtilesextractor
+
+import android.Manifest
+import android.content.Context
+import android.content.pm.PackageManager
+import android.support.v4.content.ContextCompat
+
+object Utils {
+    fun hasPermissions(context: Context): Boolean {
+        if (ContextCompat.checkSelfPermission(context, Manifest.permission.READ_EXTERNAL_STORAGE)
+                != PackageManager.PERMISSION_GRANTED)
+            return false
+        if (ContextCompat.checkSelfPermission(context, Manifest.permission.WRITE_EXTERNAL_STORAGE)
+                != PackageManager.PERMISSION_GRANTED)
+            return false
+        return true
+    }
+}