+++ CHANGELOG.md
@@ -1,4 +1,11 @@
+# Changelog
+
+## 0.0.2
+
+* Added asynchronous working so the extraction doesn't freeze the UI while working.
+
+* Updated the example
+
 ## 0.0.1
 
-* First release with basic functionality for extracting
-the tiles from a .mbtiles file.
+* First release with basic functionality for extracting the tiles from a .mbtiles file.